components/store_pdb_component/requirements.txt
pyarrow==15.0.0
google-cloud-storage==2.15.0
python-dotenv==1.0.1